🚚 rename hooks to kebab-case
