+++ packages/kitchn/src/components/Modal/index.tsx
@@ -3,9 +3,9 @@ import { createPortal } from "react-dom";
 import styled from "styled-components";
 
 import { withDecorator } from "../../hoc/withDecorator";
-import { useBreakpoint } from "../../hooks/useBreakpoint";
-import { useKeyboard } from "../../hooks/useKeyboard";
-import { usePortal } from "../../hooks/usePortal";
+import { useBreakpoint } from "../../hooks/use-breakpoint";
+import { useKeyboard } from "../../hooks/use-keyboard";
+import { usePortal } from "../../hooks/use-portal";
 import { KitchnComponent } from "../../types";
 import { fadeInDown, fadeOutUp } from "../../utils/animate";
 import { KeyCode } from "../../utils/codes";
