+++ packages/kitchn/src/components/NavigationMenu/Item/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 
 import { NavigationMenuItemContext } from "../../../contexts/NavigationMenuItem";
-import useNavigationMenu from "../../../hooks/useNavigationMenu";
+import useNavigationMenu from "../../../hooks/use-navigation-menu";
 import { getId } from "../../../utils";
 import NavigationMenuContent from "../Content";
