+++ packages/kitchn/src/hooks/index.ts
@@ -1,19 +1,19 @@
-export * from "./useBreakpoint";
-export * from "./useCheckbox";
-export * from "./useClickAnywhere";
-export * from "./useClickAway";
-export * from "./useClipboard";
-export * from "./useCombobox";
-export * from "./useCurrentState";
-export * from "./useDOMObserver";
-export * from "./useKeyboard";
-export * from "./useMediaQuery";
-export * from "./useModal";
-export * from "./usePortal";
-export * from "./usePrevious";
-export * from "./useRect";
-export * from "./useResize";
-export * from "./useSSR";
-export * from "./useTheme";
-export * from "./useToasts";
-export * from "./useWarning";
+export * from "./use-breakpoint";
+export * from "./use-checkbox";
+export * from "./use-click-anywhere";
+export * from "./use-click-away";
+export * from "./use-clipboard";
+export * from "./use-combobox";
+export * from "./use-current-state";
+export * from "./use-domobserver";
+export * from "./use-keyboard";
+export * from "./use-media-query";
+export * from "./use-modal";
+export * from "./use-portal";
+export * from "./use-previous";
+export * from "./use-rect";
+export * from "./use-resize";
+export * from "./use-ssr";
+export * from "./use-theme";
+export * from "./use-toasts";
+export * from "./use-warning";
