+++ packages/kitchn/src/hooks/use-breakpoint.ts
@@ -1,6 +1,6 @@
 import { useTheme } from "styled-components";
 
-import { useMediaQuery } from "./useMediaQuery";
+import { useMediaQuery } from "./use-media-query";
 
 export type UseBreakpointResult = {
   isMobile: boolean;
