+++ packages/kitchn/src/native/hooks/index.ts
@@ -1,1 +1,1 @@
-export * from "./useTheme";
+export * from "./use-theme";
